plumbing: format/idxfile, Compute name size in int
`readObjectNames` computed the per-bucket name-buffer length as
`int(buckets * idSize)`, performing the multiplication in `uint32`
before widening. With SHA-256 (`idSize` 32) the multiplication wraps
once `buckets` exceeds 2^27, silently producing a smaller-than-needed
buffer; a subsequent `ReadFull` would then desynchronize the input
stream from the decoder's expectations and surface as an opaque
short-read or checksum mismatch later.

Promote both operands to `int` before multiplying, and apply the
same shape to the `Offset32`/`CRC32` allocations for consistency.
With the fanout cap added in the previous commit the wrap is no
longer reachable in practice, but typing the arithmetic correctly
removes the trap if the cap is ever loosened.

+++ plumbing/format/idxfile/decoder.go
@@ -141,7 +141,7 @@ func readFanout(idx *MemoryIndex, r io.Reader) error {
 }
 
 func readObjectNames(idx *MemoryIndex, r io.Reader) error {
-	idSize := uint32(idx.idSize())
+	idSize := idx.idSize()
 
 	for k := range fanout {
 		var buckets uint32
@@ -157,15 +157,15 @@ func readObjectNames(idx *MemoryIndex, r io.Reader) error {
 
 		idx.FanoutMapping[k] = len(idx.Names)
 
-		nameLen := int(buckets * idSize)
+		nameLen := int(buckets) * idSize
 		bin := make([]byte, nameLen)
 		if _, err := io.ReadFull(r, bin); err != nil {
 			return err
 		}
 
 		idx.Names = append(idx.Names, bin)
-		idx.Offset32 = append(idx.Offset32, make([]byte, buckets*4))
-		idx.CRC32 = append(idx.CRC32, make([]byte, buckets*4))
+		idx.Offset32 = append(idx.Offset32, make([]byte, int(buckets)*4))
+		idx.CRC32 = append(idx.CRC32, make([]byte, int(buckets)*4))
 	}
 
 	return nil
